# docs/api/openapi-generated.yaml
components:
  schemas:
...
    CreateWebhookRequest:
      description: Create webhook request body.
      properties:
        eventTypes:
          items:
            type: string
          type: array
        name:
          type: string
        secret:
          type:
          - string
          - 'null'
        url:
          type: string
      required:
      - name
      - url
      - eventTypes
      type: object
...
    ListWebhookLogsResponse:
      description: List webhook logs response.
      properties:
        data:
          items:
            $ref: '#/components/schemas/WebhookLogResponse'
          type: array
        nextCursor:
          type:
          - string
          - 'null'
      required:
      - data
      type: object
    ListWebhooksResponse:
      description: List webhooks response.
      properties:
        data:
          items:
            $ref: '#/components/schemas/WebhookResponse'
          type: array
      required:
      - data
      type: object
...
    ProblemDetails:
      allOf:
      - description: RFC 7807 extension fields (e.g. for 409 `StaleConversation` extra data).
        type:
        - object
        - 'null'
      - properties:
          detail:
            description: Detailed error description.
            type:
            - string
            - 'null'
          errors:
            description: Field-level validation errors.
            items:
              $ref: '#/components/schemas/ValidationError'
            type:
            - array
            - 'null'
          instance:
            description: URI of the specific resource where the error occurred.
            type:
            - string
            - 'null'
          status:
            description: HTTP status code.
            format: int32
            minimum: 0
            type: integer
          title:
            description: Human-readable error summary.
            type: string
          type:
            description: Error type URI for programmatic identification.
            type: string
        required:
        - type
        - title
        - status
        type: object
      description: RFC 7807 Problem Details error response.
...
    TestWebhookResponse:
      description: Test webhook response.
      properties:
        eventLogId:
          format: uuid
          type: string
        status:
          type: string
      required:
      - eventLogId
      - status
      type: object
...
    UpdateWebhookRequest:
      description: Update webhook request body.
      properties:
        enabled:
          type:
          - boolean
          - 'null'
        eventTypes:
          items:
            type: string
          type:
          - array
          - 'null'
        name:
          type:
          - string
          - 'null'
        secret:
          type:
          - string
          - 'null'
        url:
          type:
          - string
          - 'null'
      type: object
...
    ValidationError:
      description: A single field validation error.
      properties:
        code:
          description: Error code for programmatic handling (e.g. "required", "`too_long`").
          type: string
        field:
          description: Field path (e.g. "name", "config.apiKey").
          type: string
        message:
          description: Human-readable error message.
          type: string
      required:
      - field
      - message
      - code
      type: object
...
    WebhookLogResponse:
      description: Webhook event log response item.
      properties:
        attempts:
          format: int32
          type: integer
        completedAt:
          type:
          - string
          - 'null'
        createdAt:
          type: string
        eventType:
          type: string
        id:
          format: uuid
          type: string
        maxAttempts:
          format: int32
          type: integer
        payload: {}
        responseStatus:
          format: int32
          type:
          - integer
          - 'null'
        status:
          type: string
        webhookId:
          format: uuid
          type: string
      required:
      - id
      - webhookId
      - eventType
      - payload
      - status
      - attempts
      - maxAttempts
      - createdAt
      type: object
    WebhookResponse:
      description: Webhook response item.
      properties:
        createdAt:
          type: string
        createdBy:
          format: uuid
          type: string
        enabled:
          type: boolean
        eventTypes: {}
        hasSecret:
          type: boolean
        id:
          format: uuid
          type: string
        name:
          type: string
        projectId:
          format: uuid
          type: string
        updatedAt:
          type: string
        url:
          type: string
      required:
      - id
      - projectId
      - name
      - url
      - hasSecret
      - eventTypes
      - enabled
      - createdBy
      - createdAt
      - updatedAt
      type: object
  securitySchemes:
    bearer_auth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: AI-assisted conference/event project management system
  license:
    name: ''
  title: Conf-Ops API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/projects/{projectId}/webhooks:
    get:
      description: |-
        # Errors

        Returns `ProblemDetails` on database failure.
      operationId: list_webhooks
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWebhooksResponse'
          description: Webhook list
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: List webhooks for a project.
      tags:
      - webhooks
    post:
      description: |-
        # Errors

        Returns `ProblemDetails` on validation or database failure.
      operationId: create_webhook
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
          description: Webhook created
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Create a new webhook.
      tags:
      - webhooks
  /api/v1/projects/{projectId}/webhooks/{webhookId}:
    delete:
      description: |-
        # Errors

        Returns `ProblemDetails` on not found.
      operationId: delete_webhook
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: Webhook deleted
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Delete a webhook.
      tags:
      - webhooks
    get:
      description: |-
        # Errors

        Returns `ProblemDetails` on not found.
      operationId: get_webhook
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
          description: Webhook details
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Get a webhook by ID.
      tags:
      - webhooks
    put:
      description: |-
        # Errors

        Returns `ProblemDetails` on validation or not found.
      operationId: update_webhook
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
          description: Webhook updated
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Update a webhook.
      tags:
      - webhooks
  /api/v1/projects/{projectId}/webhooks/{webhookId}/logs:
    get:
      description: |-
        # Errors

        Returns `ProblemDetails` on database failure.
      operationId: list_webhook_logs
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        schema:
          format: uuid
          type: string
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          format: uuid
          type: string
      - description: Max items to return
        in: query
        name: limit
        required: false
        schema:
          format: int64
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWebhookLogsResponse'
          description: Webhook event logs
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: List event logs for a webhook.
      tags:
      - webhooks
  /api/v1/projects/{projectId}/webhooks/{webhookId}/test:
    post:
      description: |-
        # Errors

        Returns `ProblemDetails` on failure.
      operationId: test_webhook
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestWebhookResponse'
          description: Test event sent
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Send a test event to a webhook.
      tags:
      - webhooks
...